# src/main/java/com/api/TCCnnect/controller/Auth/TokenController.java
package com.api.TCCnnect.controller.Auth;

import com.api.TCCnnect.dto.DadosAutenticacaoDTO;
import com.api.TCCnnect.dto.DadosTokenJwtDTO;
import com.api.TCCnnect.model.User;
import com.api.TCCnnect.services.TokenService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/auth")
public class TokenController {

    private final AuthenticationManager manager;
    private  final TokenService tokenService;

    public TokenController(AuthenticationManager manager, TokenService tokenService) {
        this.manager = manager;
        this.tokenService = tokenService;
    }

    @PostMapping("/login")
    public ResponseEntity<?> efetuarLogin(@RequestBody @Valid DadosAutenticacaoDTO dados) {
        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(dados.login(), dados.password());
        Authentication authentication = manager.authenticate(authenticationToken);
        String tokenJWT = tokenService.gerarToken((User) authentication.getPrincipal());
        return ResponseEntity.ok(new DadosTokenJwtDTO(tokenJWT));

    }
}