# swagger.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCCnnect API",
    "description": "Documentação da API do TCCnnect",
    "contact": {
      "name": "Equipe TCCnnect",
      "email": "contato@tccnnect.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "jwt_auth": []
    }
  ],
  "paths": {
    "/users/me": {
      "put": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "updateUserProfile",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDTO"
                }
              }
            }
          }
        }
      }
    },
    "/follow": {
      "post": {
        "tags": [
          "follow-controller"
        ],
        "operationId": "followUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "Cria um novo usuário",
        "description": "Endpoint para registrar um novo usuário na aplicação.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignUpDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuário criado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "token-controller"
        ],
        "operationId": "efetuarLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DadosAutenticacaoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "getUserProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDTO"
                }
              }
            }
          }
        }
      }
    },
    "/users/search": {
      "get": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "searchUserProfile",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/follow/following": {
      "get": {
        "tags": [
          "follow-controller"
        ],
        "operationId": "getFollowing",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowingResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "summary": "Obtém o perfil do usuário logado",
        "description": "Retorna os dados do perfil do usuário autenticado com base no token JWT.",
        "operationId": "getMyProfile",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil do usuário retornado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDTO"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou não fornecido",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/users/delete/me": {
      "delete": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "deleteUserProfile",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserProfileDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "login": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string"
          }
        }
      },
      "FollowRequestDTO": {
        "type": "object",
        "properties": {
          "followerId": {
            "type": "string",
            "format": "uuid"
          },
          "followedId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FollowResponseDTO": {
        "type": "object",
        "properties": {
          "followerName": {
            "type": "string"
          },
          "followedName": {
            "type": "string"
          }
        }
      },
      "UserSignUpDTO": {
        "type": "object",
        "description": "Representa os dados necessários para o cadastro de um novo usuário",
        "properties": {
          "login": {
            "type": "string",
            "description": "E-mail do usuário que será usado como login",
            "example": "usuario@exemplo.com",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "senha123",
            "minLength": 1
          }
        }
      },
      "DadosAutenticacaoDTO": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "FollowingResponseDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "jwt_auth": {
        "type": "http",
        "name": "jwt_auth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}